+++ packages/atproto/test/src/sync/sync_service_test.dart
@@ -704,4 +704,72 @@ void main() {
       }
     }, timeout: Timeout(Duration(minutes: 2)));
   });
+
+  group('.notifyCrawlingServiceOfUpdate', () {
+    test('normal case', () async {
+      final sync = SyncService(
+        did: 'test',
+        protocol: core.Protocol.https,
+        service: 'test',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: atp_test.createMockedPostClient(
+          'test/src/sync/data/notify_crawling_service_of_update.json',
+        ),
+      );
+
+      final response = await sync.notifyCrawlingServiceOfUpdate(
+        hostname: 'bsky.social',
+      );
+
+      expect(response, isA<core.XRPCResponse>());
+      expect(response.data, isA<core.EmptyData>());
+    });
+
+    test('when unauthorized', () async {
+      final sync = SyncService(
+        did: 'test',
+        protocol: core.Protocol.https,
+        service: 'test',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: atp_test.createMockedPostClient(
+          'test/src/data/error.json',
+          statusCode: 401,
+        ),
+      );
+
+      atp_test.expectUnauthorizedException(
+        () async => await sync.notifyCrawlingServiceOfUpdate(
+          hostname: 'bsky.social',
+        ),
+      );
+    });
+
+    test('when rate limit exceeded', () async {
+      final sync = SyncService(
+        did: 'test',
+        protocol: core.Protocol.https,
+        service: 'test',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: atp_test.createMockedPostClient(
+          'test/src/data/error.json',
+          statusCode: 429,
+        ),
+      );
+
+      atp_test.expectRateLimitExceededException(
+        () async => await sync.notifyCrawlingServiceOfUpdate(
+          hostname: 'bsky.social',
+        ),
+      );
+    });
+  });
 }